ai-worker/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.1+cpu
numpy==1.26.4

pika
open-clip-torch
Pillow
requests
scikit-learn
python-logging-loki
